Restructure admin and models organization.
Split `admin.py` and `models.py` into modular files for better organization and maintainability. Updated project configurations to reflect changes, including new Django-specific settings in `.idea/prirodou.iml`.
