+++ data/admin/__init__.py
@@ -1,0 +1,2 @@
+from data.admin.author import AuthorAdmin
+from data.admin.article import ArticleAdmin
